% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algn_rasts.R
\name{algn_rasts}
\alias{algn_rasts}
\title{Align cost rasters}
\usage{
algn_rasts(
  dtm,
  dns,
  rgh,
  force = F,
  out_file_dtm = NULL,
  out_file_dns = NULL,
  out_file_rgh = NULL,
  overwrite = F
)
}
\arguments{
\item{dtm}{SpatRaster. The coarse-scale (e.g., 10 m) DTM. This may be
the product of \code{\link[=gen_dtm]{gen_dtm()}}.}

\item{dns}{SpatRaster. The vegetation density raster dataset. This may be
the product of \code{\link[=gen_dns]{gen_dns()}}.}

\item{rgh}{SpatRaster. The ground surface roughness raster dataset. This may
be the product of \code{\link[=gen_rgh]{gen_rgh()}}.}

\item{force}{Boolean. If spatial incongruence exists between any of the three
input rasters, you can optionally choose to force them into congruence.}

\item{out_file_dtm}{character. Optionally, if \code{force == T}, you can save the
new, congruent DTM to disk by supplying a file path with a .tif extension.}

\item{out_file_dns}{character. Optionally, if \code{force == T}, you can save the
new, congruent vegetation density raster to disk by supplying a file path
with a .tif extension.}

\item{out_file_rgh}{character. Optionally, if \code{force == T}, you can save the
new, congruent ground surface roughness raster to disk by supplying a file
path with a .tif extension.}

\item{overwrite}{Boolean. If \code{force == T} and you opt to write any or all of
your new, congruent rasters to disk, this parameter specifies if you are
willing to overwrite existing files.}
}
\value{
A list with three \code{SpatRaster} items:
\itemize{
\item \code{dtm} is either the original or new, congruence-forced DTM.
\item \code{dns} is either the original or new, congruence-forced vegetation density
raster.
\item \code{rgh} is either the original or new, congruence-forced ground surface
roughness raster.
}
}
\description{
\code{algn_rasts()} checks to see if all three of your input cost rasters (DTM,
vegetation density, and ground surface roughness) are spatially congruent
(crs, resolution, extent, origin). If not, you can optionally force them into
congruence.
}
\details{
If \code{force == T}, it is important that the input rasters are at least quite
similar in crs, resolution, extent, and origin, as a relatively simple
resampling technique will be employed to force them into congruence. If they
are very different, the resulting rasters could have spatial artifacts and
may not accurately represent the landscape characteristics to a suitable
degree of accuracy. Note that the input DTM will be used as the "reference"
raster dataset to which the others will be forced to match the spatial
characteristics of.
}
