% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bld_tm.R
\name{bld_tm}
\alias{bld_tm}
\title{Build transition matrix}
\usage{
bld_tm(dtm, dns, rgh, clf, wtr)
}
\arguments{
\item{dtm}{SpatRaster. The coarse-scale (e.g., 10 m) DTM. This may be
the product of \code{\link[=gen_dtm]{gen_dtm()}}.}

\item{dns}{SpatRaster. The vegetation density raster dataset. This may be
the product of \code{\link[=gen_dns]{gen_dns()}}.}

\item{rgh}{SpatRaster. The ground surface roughness raster dataset. This may
be the product of \code{\link[=gen_rgh]{gen_rgh()}}.}

\item{clf}{SpatRaster. The cliff barrier raster dataset. This may be the
product of \code{\link[=get_bars]{get_bars()}}.}

\item{wtr}{SpatRaster. The water barrier raster dataset. This may be the
product of \code{\link[=get_bars]{get_bars()}}.}
}
\value{
A \code{transitionMatrix}, with values representing the conductance
(inverse of cost) between adjacent cells in your study area.
}
\description{
\code{bld_tm()} builds a \code{\link[gdistance:Transition-slots]{gdistance::transitionMatrix()}} using the three input
cost rasters: (1) DTM (from which directional slope will be calculated); (2)
vegetation density; and (3) ground surface roughness; and the two input barrier
rasters: (1) cliffs (slopes >= 45 deg.); and (2) waterbodies.
}
\details{
It is imperative to check for (and optionally enforce) spatial congruence
between the three input cost rasters using \code{\link[=algn_rasts]{algn_rasts()}} prior to running
this function. Provided that \code{\link[=get_bars]{get_bars()}} is run after \code{\link[=algn_rasts]{algn_rasts()}}, the
two barrier rasters will automatically be aligned and ready for use here.

Note that building transition matrices is a computationally intensive
process, requiring many pairwise calculations between neighboring cells in a
potentially large study area, and the storage of those calculations in
memory. You may consider testing the algorithm on a small study area prior to
employment on a broad spatial scale.
}
