% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{common_error_messages}
\alias{common_error_messages}
\title{Common error messages}
\usage{
common_error_messages(
  n,
  m,
  p,
  qvs,
  q,
  x,
  delta,
  ww,
  zz,
  run.NPMLEs,
  run.NPNA,
  run.NPNA_avg,
  run.NPNA_wrong,
  tval,
  tval0,
  z.use,
  w.use,
  update.qs,
  know.true.groups,
  true.groups,
  run.prediction.accuracy
)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{p}{number of populations, must be at least 2.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{q}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}

\item{x}{a numeric vector of length \code{n} containing the observed event times
for each person in the sample.}

\item{delta}{a numeric vector of length \code{n} that denotes
censoring (1 denotes event is observed, 0 denotes event is censored).}

\item{ww}{a numeric vector of length \code{n} containing the values of the continuous
covariate for each person in the sample.}

\item{zz}{a numeric vector of length \code{n} containing the values of the discrete
covariate for each person in the sample.}

\item{run.NPMLEs}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data based on the type-I nonparametric maximum likelihood
estimator. This is referred to as the "Kaplan-Meier" estimator in the reference paper, and does
not include covariates nor dynamic landmarking.}

\item{run.NPNA}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that accounts for covariates and dynamic
landmarking. This estimator is called "NPNA" in the referenced paper.}

\item{run.NPNA_avg}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that averages out over the observed covariates.
This is referred to as NPNA_marg in the referenced paper.}

\item{run.NPNA_wrong}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that ignores landmarking. This is referred to as
NPNA_{t_0=0} in the paper.}

\item{tval}{numeric vector of time points at which the distribution function is evaluated, all values must
be non-negative.}

\item{tval0}{numeric vector of time points representing the landmark times. All values must be non-negative
and smaller than the maximum of \code{tval}.}

\item{z.use}{numeric vector at which to evaluate the discrete covariate \eqn{Z} at in the estimated distribution function.
The values of \code{z.use} must be in the range of the observed \code{zz}.}

\item{w.use}{numeric vector at which to evaluate the continuous covariate \eqn{W} at in the estimated distribution function.
The values of \code{w.use} must be in the range of the observed \code{ww}.}

\item{update.qs}{logical indicator. If TRUE, the mixture proportions \code{q} will be updated. This is currently not implemented.}

\item{know.true.groups}{logical indicator. If TRUE, then we know the population identifier for each person in the sample.
This option is only used for simulation studies. Default is FALSE.}

\item{true.groups}{numeric vector of length \code{n} denoting the population identifier for each person in the sample.
Default is NULL.}

\item{run.prediction.accuracy}{logical indicator. If TRUE, then we compute the prediction accuracy measures, including the
area under the receiver operating characteristic curve (AUC) and the Brier Score (BS). Prediction accuracy is only valid
in simulation studies where \code{know.true.groups}=TRUE and \code{true.groups} is available.}
}
\value{
Error or warning messages when input is not appropriate for the methods.
}
\description{
Function to check for common errors in the input.
}
