% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.r}
\alias{compute.r}
\title{Sample sizes of mixture proportion groups.}
\usage{
compute.r(n, m, p, qvs, q.use)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{p}{number of populations, must be at least 2.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{q.use}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}
}
\description{
Computes the number of individuals in each mixture proportion group.
}
