% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.uset}
\alias{compute.uset}
\title{Mixture porportion group identifier}
\usage{
compute.uset(n, m, p, qvs, q.use)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{p}{number of populations, must be at least 2.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{q.use}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}
}
\value{
A numeric vector of length \code{n} containing the mixture proportion
group to which each person belongs.
}
\description{
Computes the mixture proportion group identifier.
}
\section{Details}{

The matrix \code{qvs} contains all mixture proportions. The function
\code{compute.uset} assigns individual \eqn{i} to mixture proportion \eqn{u},
for \eqn{u=1,\ldots,m}, if \eqn{q_i=qvs_{u}}.
}

