% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns_fun.R
\name{dns_fun}
\alias{dns_fun}
\title{Density function}
\usage{
dns_fun(z, hgt_1, hgt_2)
}
\arguments{
\item{z}{numeric. Aboveground lidar point heights.}

\item{hgt_1}{numeric. The bottom value of the vegetation density height
range.}

\item{hgt_2}{numeric. The top value of the vegetation density height range.}
}
\value{
A numeric value representing normalized relative lidar point
density.
}
\description{
\code{dns_fun()} defines the calculation of normalized relative lidar point
density (proxy for vegetation density) that is applied to every pixel within
the extent of a lidar dataset using \code{gen_dns()}.
}
