% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{error_max_value}
\alias{error_max_value}
\title{Error from exceeding maximum}
\usage{
error_max_value(x, max_x, variable_name)
}
\arguments{
\item{x}{numeric vector or array object.}

\item{max_x}{maximum allowed value in the vector or array.}

\item{variable_name}{name of object x.}
}
\value{
Error message if any component of \code{x} is larger than \code{max_x}.
}
\description{
Reports error if x > max_x
}
\examples{
## Reports an error (not run)
##love <- c(1,2)
##max_love <- 1.5
##error_max_value(love,max_love,get_variable_name(love))

}
