% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{error_nonpositive_value}
\alias{error_nonpositive_value}
\title{Check non-negativity}
\usage{
error_nonpositive_value(x, variable_name)
}
\arguments{
\item{x}{numeric vector.}

\item{variable_name}{name of object x.}
}
\value{
Error message if any component of \code{x} < 0.
}
\description{
Stops code if  \code{x} is not >=0.
}
\examples{
## Returns error (not run)
##test <- c(-1,2)
##error_nonpositive_value(test,get_variable_name(test))

}
