% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{error_number_of_predictors}
\alias{error_number_of_predictors}
\title{Check number of predictors are valid}
\usage{
error_number_of_predictors(
  number_of_functional_predictors,
  number_of_nonfunctional_predictors
)
}
\arguments{
\item{number_of_functional_predictors}{Number of functional predictors.}

\item{number_of_nonfunctional_predictors}{Number of nonfunctional predictors.}
}
\value{
Error if  number_of_functional_predictors =0 and
number_of_nonfunctional_predictors=0.
}
\description{
Reports an error if the number_of_functional_predictors =0 and
number_of_nonfunctional_predictors=0. Our model runs only if we have at least one
functional_predictor or at least one nonfunctional_predictor.
}
\examples{
## Returns error (not run)
#number_of_functional_predictors <- 0
#number_of_nonfunctional_predictors <- 0
#error_number_of_predictors(number_of_functional_predictors,number_of_nonfunctional_predictors)

}
