% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{error_positive_value}
\alias{error_positive_value}
\title{Check positivity}
\usage{
error_positive_value(x, variable_name)
}
\arguments{
\item{x}{numeric vector.}

\item{variable_name}{name of object x.}
}
\value{
Error message if any component of \code{x} is <= 0.
}
\description{
Stops code if  \code{x} is not >0.
}
\examples{
## Returns error (not run)
#test <- c(-1,2)
#error_positive_value(test,get_variable_name(test))

}
