% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{error_type_number}
\alias{error_type_number}
\title{Incorrect size or type}
\usage{
error_type_number(type, number, type_options)
}
\arguments{
\item{type}{character vector or array.}

\item{number}{required length of \code{type} if \code{type} is a vector, or
required dimension of \code{type} if \code{type} is an array.}

\item{type_options}{character of what \code{type} is allowed to be.}
}
\value{
Error message if any component of \code{type} is not the correct size or
allowed type.
}
\description{
Reports error if \code{type} is not the correct size specified by \code{number}
or if \code{type} is not one of the \code{type_options}.
}
\examples{
## Reports an error because of wrong type. (not run)
#number_of_functional_predictors <- 3
#functional_coefficients_basis_type <- rep("cubic-spline",number_of_functional_predictors)
#functional_basis_options <- c("fpca","quantlet")
#error_type_number(functional_coefficients_basis_type,number_of_functional_predictors,
#functional_basis_options)

}
