% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{estimator.main}
\alias{estimator.main}
\title{Estimator implementation}
\usage{
estimator.main(
  data,
  n,
  p,
  m,
  r,
  qvs,
  tval,
  tval0,
  method.label,
  z.use,
  w.use,
  update.qs,
  run.prediction.accuracy,
  do_cross_validation_AUC_BS
)
}
\arguments{
\item{data}{data matrix obtained from \code{make.data.set}}

\item{n}{sample size, must be at least 1.}

\item{p}{number of populations, must be at least 2.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{r}{numeric vector including the number of individuals in each mixture proportion
group.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{tval}{numeric vector of time points at which the distribution function is evaluated, all values must
be non-negative.}

\item{tval0}{numeric vector of time points representing the landmark times. All values must be non-negative
and smaller than the maximum of \code{tval}.}

\item{method.label}{character vector of methods implemented. This is the result from \code{get_method_label()}/}

\item{z.use}{numeric vector at which to evaluate the discrete covariate \eqn{Z} at in the estimated distribution function.
The values of \code{z.use} must be in the range of the observed \code{zz}.}

\item{w.use}{numeric vector at which to evaluate the continuous covariate \eqn{W} at in the estimated distribution function.
The values of \code{w.use} must be in the range of the observed \code{ww}.}

\item{update.qs}{logical indicator. If TRUE, the mixture proportions \code{q} will be updated. This is currently not implemented.}

\item{run.prediction.accuracy}{logical indicator. If TRUE, then we compute the prediction accuracy measures, including the
area under the receiver operating characteristic curve (AUC) and the Brier Score (BS). Prediction accuracy is only valid
in simulation studies where \code{know.true.groups}=TRUE and \code{true.group.identifier} is available.}

\item{do_cross_validation_AUC_BS}{logical indicator. If TRUE, then we compute the prediction accuracy measures, including the
area under the receiver operating characteristic curve (AUC) and the Brier Score (BS) using cross-validation. Prediction accuracy is only valid
in simulation studies where \code{know.true.groups}=TRUE and \code{true.group.identifier} is available.}
}
\value{
\code{estimator.main} returns a list containing
\itemize{
   \item{Ft.store: }{a numeric array. When
   \code{run.prediction.accuracy} is FALSE, then the results are the
   the estimated distribution functions for all \code{p} populations.
   The dimension of the array is \# of methods by \code{length(tval)} by \code{lenth(tval0)} by
   \code{length(z.use)} (when \code{z.use} is non-NULL) by \code{length(w.use)}
   (when \code{w.use} is non-NULL) by \code{p}.

   When \code{run.prediction.accuracy} is TRUE, then the results are the
   area under the receiver operating characteristic curve (AUC) and
   Brier Score (BS) for the \code{p} populations. The dimension of the array is \# of methods by
   \code{length(tval)} by \code{length(tval0)} by 2,
   where the last dimension stores the AUC and BS results.

   Results for both the estimated distributon functions and prediction
   accuracy measures (AUC, BS) are only valid when \eqn{t\geq t_0}, so
   arrays show NA for any combination for which \eqn{t<t_0}.

   }

   \item {St.store: }{a numeric array. When
   \code{run.prediction.accuracy} is FALSE, then the results are the
   the estimated distribution functions for all \code{m} mixture proportion
   groups.
   The dimension of the array is \# of methods by \code{length(tval)} by \code{lenth(tval0)} by
   \code{length(z.use)} (when \code{z.use} is non-NULL)
   by \code{length(w.use)} (when \code{w.use} is non-NULL) by \code{m}.

   When \code{run.prediction.accuracy} is TRUE, then the results are the
   area under the receiver operating characteristic curve (AUC) and
   Brier Score (BS) for the \code{m} mixture proportion groups.
   The dimension of the array is \# of methods by
   \code{length(tval)} by \code{length(tval0)} by 2,
   where the last dimension stores the AUC and BS results.

   Results for both the estimated distributon functions and prediction
   accuracy measures (AUC, BS) are only valid when \eqn{t\geq t_0}, so
   arrays show NA for any combination for which \eqn{t<t_0}.

   }

   \item{problem: }{a numeric indicator of errors in the NPNA estimator. If NULL, no error is reported.
   Otherwise, there is an error in the computation of the NPNA estimator.}

}
}
\description{
Main function to estimate the distribution function for mixture data where
the population identifiers are unknown, but the probability of belonging
to a population is known. The distribution functions are evaluated at
time points \code{tval} and adjust for dynamic landmark prediction and one
discrete covariate (\code{zz}) and one continuous covariate (\code{ww}).
}
\section{Details}{

We estimate nonparametric distribution functions for mixture data  where
the population identifiers are unknown, and the probability of belonging
to a population is known (typically estimated with external data).
The distribution functions are evaluated at
time points \code{tval}. All estimators adjust for dynamic landmark prediction.
Dynamic landmark prediction means that the distribution function is computed knowing
that the survival time, \eqn{T}, satisfies \eqn{T >t_0}
where \eqn{t_0} are the time points in \code{tval0}. The NPNA, NPNA_avg,
and NPNA_wrog adjust for one discrete covariate (\code{zz}) and one continuous covariate (\code{ww}).
}

\references{
Garcia, T.P. and Parast, L. (2020). Dynamic landmark prediction for mixture data. Biostatistics,  doi:10.1093/biostatistics/kxz052.

Garcia, T.P., Marder, K. and Wang, Y. (2017). Statistical modeling of Huntington disease onset.
In Handbook of Clinical Neurology, vol 144, 3rd Series, editors Andrew Feigin and Karen E. Anderson.

Qing, J., Garcia, T.P., Ma, Y., Tang, M.X., Marder, K., and Wang, Y. (2014).
Combining isotonic regression and EM algorithm to predict genetic risk under monotonicity constraint.
Annals of Applied Statistics, 8(2), 1182-1208.

Wang, Y., Garcia, T.P., and Ma. Y. (2012).  Nonparametric estimation for censored mixture data with
application to the Cooperative Huntington's Observational Research Trial. Journal of the American Statistical Association,
107, 1324-1338.
}
