% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dns.R
\name{gen_dns}
\alias{gen_dns}
\title{Generate vegetation density raster}
\usage{
gen_dns(
  las_hgt,
  hgt_1 = 0.85,
  hgt_2 = 1.2,
  res = 10,
  ncores = 1L,
  out_file = NULL,
  overwrite = F
)
}
\arguments{
\item{las_hgt}{An object of class \code{LAS} (i.e., from \code{\link[lidR:readLAS]{lidR::readLAS()}}) or
\code{LAScatalog} (i.e., from \code{\link[lidR:readLAScatalog]{lidR::readLAScatalog()}}). Airborne lidar point
cloud data that have been normalized to the ground surface, where Z values
represent heights above the ground.}

\item{hgt_1}{numeric. The bottom value of the vegetation density height
range. Defaults to 0.85 m, based on the work of Campbell et al. (in review).}

\item{hgt_2}{numeric. The top value of the vegetation density height range.
Defaults to 1.20 m, based on the work of Campbell et al. (in review).}

\item{res}{numeric. The size of a grid cell in point cloud coordinate units.}

\item{ncores}{numeric. If you supply a \code{LAScatalog} with multiple tiles of
lidar data, you can leverage multiple CPUs to process your data in parallel,
using \code{\link[future:plan]{future::plan()}}.}

\item{out_file}{character. Optionally, you can save the vegetation density
raster to disk by supplying a file path with a .tif extension.}

\item{overwrite}{Boolean. If you opt to write the DTM to an output file, this
parameter specifies if you are willing to overwrite an existing file.}
}
\value{
\code{SpatRaster} where each pixel represents an approximation of the
vegetation density between \code{hgt_1} and \code{hgt_2}.
}
\description{
\code{gen_dns()} calculates an approximation of vegetation density, based on
lidar point density, within a particular height range of interest.
}
\details{
To use this dataset in conjunction with a DTM and ground surface roughness
raster for the purposes of modeling least-cost paths, it is important to
ensure that the spatial resolution is consistent between all of these inputs.
}
