% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{GenerateData}
\alias{GenerateData}
\title{Generate data}
\usage{
GenerateData(n, p, m, qvs, censoring.rate, simu.setting)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{p}{number of populations, must be at least 2.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{censoring.rate}{a scalar indicating the censoring proportion. Options are 0 or 40.}

\item{simu.setting}{Character indicating simulation setting.
Options are "Log-Normal-No-Covariates", "Log-Normal-With-Covariates", "HD-No-Covariates","HD-With-Covariates".
Setting "Log-Normal-No-Covariates" and "Log-Normal-With-Covariates" refer to simulation setting 1 in Garcia and Parast (2020).
"Log-Normal-No-Covariates" means the
survival outcomes do NOT depend on the covariates, and "Log-Normal-With-Covariates" means the
survival outcomes do depend on the covariates.
Setting "HD-No-Covariates" and "HD-With-Covariates" refer to Simulation setting 2 in Garcia and Parast (2020), "HD-No-Covariates" means the
survival outcomes do NOT depend on the covariates, and "HD-With-Covariates" means the
survival outcomes do depend on the covariates.}
}
\value{
Returns a list containing
\itemize{
   \item{x: }{a numeric vector of length \code{n} containing the observed event times
for each person in the sample.}
   \item{delta: }{a numeric vector of length \code{n} that denotes
censoring (1 denotes event is observed, 0 denotes event is censored).}
   \item{q: }{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}
   \item{ww: }{a numeric vector of length \code{n} containing the values of the continuous
covariate for each person in the sample.}
   \item{zz: }{ a numeric vector of length \code{n} containing the values of the discrete
covariate for each person in the sample.}
   \item{true.group.identifier: }{numeric vector of length \code{n} denoting the population identifier for each person in the sample.
   This is only used simulation study to evaluate the prediction accuracy of our methods.}
}
}
\description{
Produces data from different populations with the probability of
belonging to a population. Also produces one discrete covariate and one continuous covariate.
}
\references{
Garcia, T.P. and Parast, L. (2020). Dynamic landmark prediction for mixture data. Biostatistics,  doi:10.1093/biostatistics/kxz052.
}
