% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.empty.list}
\alias{get.empty.list}
\title{Create empty list}
\usage{
get.empty.list(names)
}
\arguments{
\item{names}{character vector of names for the list}
}
\description{
Returns an empty list where list names are determined by \code{names}.
}
\examples{
list_names <- c("kitty","dog","bird")
get.empty.list(list_names)

}
