% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bars.R
\name{get_bars}
\alias{get_bars}
\title{Get barriers}
\usage{
get_bars(
  dtm,
  in_wtr = NULL,
  out_file_clf = NULL,
  out_file_wtr = NULL,
  overwrite = F
)
}
\arguments{
\item{dtm}{SpatRaster. The coarse-scale (e.g., 10 m) DTM. This may be
the product of \code{\link[=gen_dtm]{gen_dtm()}}.}

\item{in_wtr}{SpatVector. A vector polygon dataset representing area
waterbodies that will act as barriers to movement. If not supplied, in the
US, NHD data will be downloaded and used instead.}

\item{out_file_clf}{character. Optionally, if \code{force == T}, you can save the
cliff barrier raster to disk by supplying a file path with a .tif extension.}

\item{out_file_wtr}{character. Optionally, if \code{force == T}, you can save the
water barrier raster to disk by supplying a file path with a .tif extension.}

\item{overwrite}{Boolean. If you opt to write either or both of your barrier
rasters to disk, this parameter specifies if you are willing to overwrite
existing files.}
}
\value{
A list with two \code{SpatRaster} items:
\itemize{
\item \code{clf} is the cliff barrier raster.
\item \code{wtr} is the water barrier raster.
}
}
\description{
\code{get_bars()} allows you to create two raster datasets that represent barriers
on the landscape that would represent impassible pedestrian obstacles. The
two barriers used in STRIDE are cliffs (slope >= 45 deg.) and area
waterbodies (e.g., ponds, lakes). The cliffs are derived from the coarse-
scale (e.g., 10 m) DTM produced by \code{\link[=gen_dtm]{gen_dtm()}}, and the waterbodies can
either be supplied as polygon vector file or downloaded using the \code{nhdR}
package. Note that National Hydrography Dataset (NHD) data are only available
in the US.
}
\details{
This function should be run after \code{\link[=algn_rasts]{algn_rasts()}} in the STRIDE modeling
workflow. This is because the cliff and water barrier rasters that are
produced by this function will be spatially aligned (crs, resolution, extent,
and origin) to the input DTM. If the DTM already aligns with the vegetation
density and ground surface roughness rasters (i.e., \code{\link[=algn_rasts]{algn_rasts()}} ran
successfully), then the new barrier rasters will likewise align, which is
imperative for successful STRIDE modeling.
}
