% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_method_label}
\alias{get_method_label}
\title{Method names}
\usage{
get_method_label(
  run.NPMLEs,
  run.NPNA,
  run.NPNA_avg,
  run.NPNA_wrong,
  run.OLS,
  run.WLS,
  run.EFF,
  run.EMPAVA
)
}
\arguments{
\item{run.NPMLEs}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data based on the type-I and type II nonparametric maximum likelihood
estimators. The type I nonparametric maximum likelihood estimator is referred
to as the "Kaplan-Meier" estimator in Garcia and Parast (2020). Neither the type I nor type II
adjust for covariates.}

\item{run.NPNA}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that accounts for covariates and dynamic
landmarking. This estimator is called "NPNA" in Garcia and Parast (2020).}

\item{run.NPNA_avg}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that averages out over the observed covariates.
This is referred to as NPNA_marg in Garcia and Parast (2020).}

\item{run.NPNA_wrong}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that adjusts for covariates, but
ignores landmarking. This is referred to as NPNA_{t_0=0} in Garcia and Parast (2020).}

\item{run.OLS}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using an ordinary least squares influence function. The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.WLS}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using a weighted least squares influence function. The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.EFF}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using the efficient influence function based on Hilbert space projection theory results.
The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.EMPAVA}{logical indicator. If TRUE, we compute the distribution function for the mixture data
based on an expectation-maximization (EM) algorithm that uses the pool adjacent violators algorithm (PAVA)
from isotone regression to yield a non-negative and monotone estimator. This estimator does not adjust
for covariates.  See details in Qing et al (2014).}
}
\value{
A character vector of method names. If \code{run.NPMLEs} is TRUE,
the character vector includes "NPMLE1" and "NPMLE2". If \code{run.NPNA} is TRUE,
the character vector includes "NPNA". If \code{run.NPNA_avg} is TRUE,
the character vector includes "NPNA_avg". If \code{run.NPNA_wrong} is TRUE,
the character vector includes "NPNA_wrong". If \code{run.OLS} is TRUE,
the character vector includes "OLSIPW","OLSAIPW","OLSIMP". If \code{run.WLS} is TRUE,
the character vector includes "WLSIPW","WLSAIPW","WLSIMP". If \code{run.EFF} is TRUE,
the character vector includes "EFFIPW","EFFAIPW","EFFIMP".
}
\description{
Creates character vector of method names.
}
\references{
Garcia, T.P. and Parast, L. (2020). Dynamic landmark prediction for mixture data. Biostatistics,  doi:10.1093/biostatistics/kxz052.

Garcia, T.P., Marder, K. and Wang, Y. (2017). Statistical modeling of Huntington disease onset.
In Handbook of Clinical Neurology, vol 144, 3rd Series, editors Andrew Feigin and Karen E. Anderson.

Qing, J., Garcia, T.P., Ma, Y., Tang, M.X., Marder, K., and Wang, Y. (2014).
Combining isotonic regression and EM algorithm to predict genetic risk under monotonicity constraint.
Annals of Applied Statistics, 8(2), 1182-1208.

Wang, Y., Garcia, T.P., and Ma. Y. (2012).  Nonparametric estimation for censored mixture data with
application to the Cooperative Huntington's Observational Research Trial. Journal of the American Statistical Association,
107, 1324-1338.
}
