% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_variable_name}
\alias{get_variable_name}
\title{Variable name}
\usage{
get_variable_name(variable)
}
\arguments{
\item{variable}{any object type.}
}
\value{
Character name of variable.
}
\description{
Reports variable name.
}
\examples{
love <- c(1,2)
get_variable_name(love)

dog <- matrix(1:4,2,2)
get_variable_name(dog)

}
