% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make.data.set}
\alias{make.data.set}
\title{Data matrix for dynamic landmark prediction of mixture data}
\usage{
make.data.set(
  n,
  m,
  p,
  qvs,
  q,
  x,
  delta,
  ww,
  zz,
  know.true.groups,
  true.group.identifier
)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{p}{number of populations, must be at least 2.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{q}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}

\item{x}{a numeric vector of length \code{n} containing the observed event times
for each person in the sample.}

\item{delta}{a numeric vector of length \code{n} that denotes
censoring (1 denotes event is observed, 0 denotes event is censored).}

\item{ww}{a numeric vector of length \code{n} containing the values of the continuous
covariate for each person in the sample.  Can be NULL.}

\item{zz}{a numeric vector of length \code{n} containing the values of the discrete
covariate for each person in the sample.  Can be NULL.}

\item{know.true.groups}{logical indicator. If TRUE, then we know the population identifier for each person in the sample.
This option is only used for simulation studies. Default is FALSE.}

\item{true.group.identifier}{numeric vector of length \code{n} denoting the population identifier for each person in the sample.
Default is NULL.}
}
\value{
a matrix of the concatenated data. The columns are "x", "delta", "q1" to "qp" (correspond to the mixture proportions),
"w" (for \code{ww} if non-NULL), "z" (for \code{zz} if non-NULL),
"uset" (for "u-set" meaning the uth mixture proportion group), and
"group" (for the true population group as defined by \code{true.group.identifier}). If \code{know.true.groups} is
FALSE, then the column "group" is a vector of 0's.
}
\description{
Create data set based on user input.
}
