% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_lcp.R
\name{map_lcp}
\alias{map_lcp}
\title{Map least-cost paths}
\usage{
map_lcp(tm, pt_orig, pt_dest, out_file = NULL, overwrite = F)
}
\arguments{
\item{tm}{transitionMatrix. The transition matrix from \code{\link[=bld_tm]{bld_tm()}}.}

\item{pt_orig}{SpatVector. The origin point(s) of the least-cost path(s).}

\item{pt_dest}{SpatVector. The destination point(s) of the least-cost
path(s).}

\item{out_file}{character. Optionally, you can save the least-cost path
vector polyline to disk by supplying a file path with a .shp extension.}

\item{overwrite}{Boolean. If you opt to write the LCP to an output file, this
parameter specifies if you are willing to overwrite an existing file.}
}
\value{
A polyline \code{SpatVector} representing the least-cost path(s) between
origin(s) and destination(s).
}
\description{
\code{map_lcp()} uses \code{\link[gdistance:shortestPath]{gdistance::shortestPath()}} and \code{\link[gdistance:costDistance-methods]{gdistance::costDistance()}}
to map the least-cost path between two or more point pairs and calculates the
total travel time in seconds for each path.
}
\details{
The coordinate reference system (CRS) of \code{pt_orig} and \code{pt_dest} must match
that of \code{tm} for this function to run successfully.
}
