% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{myquantiles.hi}
\alias{myquantiles.hi}
\title{One minus alpha quantile}
\usage{
myquantiles.hi(x)
}
\arguments{
\item{x}{numeric vector of values.}
}
\value{
\code{myquantiles.hi} returns the \eqn{1-\alpha/2}-quantile.
}
\description{
Computes the \eqn{1-\alpha} quantile of a vector \code{x}.
}
