% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{myquantiles.lo}
\alias{myquantiles.lo}
\title{Alpha quantile}
\usage{
myquantiles.lo(x)
}
\arguments{
\item{x}{numeric vector of values.}
}
\value{
\code{myquantiles.lo} returns the \eqn{\alpha/2}-quantile.
}
\description{
Computes the alpha-quantile of a vector \code{x}.
}
