% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_hgt.R
\name{norm_hgt}
\alias{norm_hgt}
\title{Normalize lidar point cloud heights relative to ground surface}
\usage{
norm_hgt(
  las,
  algorithm = lidR::tin(),
  ncores = 1L,
  out_path = NULL,
  out_name = NULL
)
}
\arguments{
\item{las}{An object of class \code{LAS} (i.e., from \code{\link[lidR:readLAS]{lidR::readLAS()}}) or
\code{LAScatalog} (i.e., from \code{\link[lidR:readLAScatalog]{lidR::readLAScatalog()}}). Airborne lidar point
cloud data in its raw format, where Z values represent elevations above mean
sea level or some reference ellipsoid.}

\item{algorithm}{function. A function that implements an algorithm to
interpolate a ground surface, the elevation of which will be subtracted from
each lidar point to compute height. \code{lidR} implements \code{\link[lidR:dtm_idw]{lidR::knnidw()}},
\code{\link[lidR:dtm_tin]{lidR::tin()}}, and \code{\link[lidR:dtm_kriging]{lidR::kriging()}} (see respective documentation and
examples).}

\item{ncores}{numeric. If you supply a \code{LAScatalog} with multiple tiles of
lidar data, you can leverage multiple CPUs to process your data in parallel,
using \code{\link[future:plan]{future::plan()}}.}

\item{out_path}{character. Optionally, you can save the height-normalized
lidar tile(s) to file. If you supply a \code{LAScatalog}, the files will inherit
their names from the input tiles, so the outputs cannot be placed in the same
directory as the input files. If you supply a \code{LAS}, you must also provide an
\code{out_name}.}

\item{out_name}{character. If you supply a \code{LAS} and specify an \code{out_path},
indicating your desire to save the output file to disk, this is the output
file name, with the extension .las or .laz, and will be stored in \code{out_path}.}
}
\value{
\code{LAS} or \code{LAScatalog} point cloud data, where point Z values will
be replaced with aboveground heights.
}
\description{
\code{norm_hgt()} normalizes lidar point cloud heights relative to the ground
surface, effectively converting elevations (i.e., above mean sea level) to
heights (i.e., above the ground level). This is a necessary prerequisite step
to calculating vegetation density.
}
