% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stop_length_message}
\alias{stop_length_message}
\title{Check dimension requirement}
\usage{
stop_length_message(x, y)
}
\arguments{
\item{x}{vector that should be length \code{y}, or an array that should have
dimensions specified by \code{y}.}

\item{y}{numeric scalar indicating required length of x,
or vector of scalars indicating required dimensions of x.}
}
\value{
Error message if \code{x} does not have dimensions specified by \code{y},
otherwise, nothing is returned.
}
\description{
Stops code if dimension of \code{x} is not as specified by \code{y}.
}
\examples{
## Returns error (not run)
#x <- c(1,2,3)
#y <- 4
#stop_length_message(x,y)


##Returns error (not run)
#x <- matrix(1:4,2,2)
#y <- c(1,2)
#stop_length_message(x,y)

}
