% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stop_options_message}
\alias{stop_options_message}
\title{Incorrect type}
\usage{
stop_options_message(x, y)
}
\arguments{
\item{x}{character vector or array that must be of type \code{y}.}

\item{y}{character vector of allowed types.}
}
\value{
Error message if any component of \code{x} is not of type \code{y}.
}
\description{
Reports an error if \code{x} is not of type \code{y}.
}
\examples{
## Reports an error (not run)
#x <- c("cat","dog")
#y <- c("cat","bird","fish")

#stop_options_message(x,y)

}
