% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stride.bootstrap.variance}
\alias{stride.bootstrap.variance}
\title{Bootstrap variance estimator}
\usage{
stride.bootstrap.variance(
  nboot,
  n,
  m,
  p,
  qvs,
  q,
  x,
  delta,
  ww,
  zz,
  run.NPMLEs,
  run.NPNA,
  run.NPNA_avg,
  run.NPNA_wrong,
  run.OLS,
  run.WLS,
  run.EFF,
  run.EMPAVA,
  tval,
  tval0,
  z.use,
  w.use,
  update.qs,
  know.true.groups,
  true.group.identifier,
  estimator_Ft,
  estimator_St,
  AUC_BS_Ft,
  AUC_BS_St,
  run.prediction.accuracy,
  do_cross_validation_AUC_BS
)
}
\arguments{
\item{nboot}{number of bootstrap replicates, must be at least 1.}

\item{n}{sample size, must be at least 1.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{p}{number of populations, must be at least 2.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{q}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}

\item{x}{a numeric vector of length \code{n} containing the observed event times
for each person in the sample.}

\item{delta}{a numeric vector of length \code{n} that denotes
censoring (1 denotes event is observed, 0 denotes event is censored).}

\item{ww}{a numeric vector of length \code{n} containing the values of the continuous
covariate for each person in the sample.  Can be NULL.}

\item{zz}{a numeric vector of length \code{n} containing the values of the discrete
covariate for each person in the sample.  Can be NULL.}

\item{run.NPMLEs}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data based on the type-I and type II nonparametric maximum likelihood
estimators. The type I nonparametric maximum likelihood estimator is referred
to as the "Kaplan-Meier" estimator in Garcia and Parast (2020). Neither the type I nor type II
adjust for covariates.}

\item{run.NPNA}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that accounts for covariates and dynamic
landmarking. This estimator is called "NPNA" in Garcia and Parast (2020).}

\item{run.NPNA_avg}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that averages out over the observed covariates.
This is referred to as NPNA_marg in Garcia and Parast (2020).}

\item{run.NPNA_wrong}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that adjusts for covariates, but
ignores landmarking. This is referred to as NPNA_{t_0=0} in Garcia and Parast (2020).}

\item{run.OLS}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using an ordinary least squares influence function. The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.WLS}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using a weighted least squares influence function. The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.EFF}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using the efficient influence function based on Hilbert space projection theory results.
The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.EMPAVA}{logical indicator. If TRUE, we compute the distribution function for the mixture data
based on an expectation-maximization (EM) algorithm that uses the pool adjacent violators algorithm (PAVA)
from isotone regression to yield a non-negative and monotone estimator. This estimator does not adjust
for covariates.  See details in Qing et al (2014).}

\item{tval}{numeric vector of time points at which the distribution function is evaluated, all values must
be non-negative.}

\item{tval0}{numeric vector of time points representing the landmark times. All values must be non-negative
and smaller than the maximum of \code{tval}.}

\item{z.use}{numeric vector at which to evaluate the discrete covariate \eqn{Z} at in the estimated distribution function.
The values of \code{z.use} must be in the range of the observed \code{zz}.  Can be NULL.}

\item{w.use}{numeric vector at which to evaluate the continuous covariate \eqn{W} at in the estimated distribution function.
The values of \code{w.use} must be in the range of the observed \code{ww}.  Can be NULL.}

\item{update.qs}{logical indicator. If TRUE, the mixture proportions \code{q} will be updated. This is currently not implemented.}

\item{know.true.groups}{logical indicator. If TRUE, then we know the population identifier for each person in the sample.
This option is only used for simulation studies. Default is FALSE.}

\item{true.group.identifier}{numeric vector of length \code{n} denoting the population identifier for each person in the sample.
Default is NULL.}

\item{estimator_Ft}{a numeric array of the estimated distribution functions for all \code{p} populations.
This is outputted from the function \code{stride.estimator}.}

\item{estimator_St}{a numeric array of the estimated distribution functions for all \code{m} mixture proportion groups.
This is output from the function \code{stride.estimator}.}

\item{AUC_BS_Ft}{a numeric array of the area under the receiver operating characteristic curve (AUC) and
Brier Score (BS) for the \code{p} populations. This is output from the function \code{stride.estimator}.}

\item{AUC_BS_St}{a numeric array of the area under the receiver operating characteristic curve (AUC) and
Brier Score (BS) for the \code{m} mixture proportion groups. This is output from
the function \code{stride.estimator}.}

\item{run.prediction.accuracy}{logical indicator. If TRUE, then we compute the prediction accuracy measures, including the
area under the receiver operating characteristic curve (AUC) and the Brier Score (BS). Prediction accuracy is only valid
in simulation studies where \code{know.true.groups}=TRUE and \code{true.group.identifier} is available.}

\item{do_cross_validation_AUC_BS}{logical indicator. If TRUE, then we compute the prediction accuracy measures, including the
area under the receiver operating characteristic curve (AUC) and the Brier Score (BS) using cross-validation. Prediction accuracy is only valid
in simulation studies where \code{know.true.groups}=TRUE and \code{true.group.identifier} is available.}
}
\value{
\code{stride.bootstrap.variance} returns a list containing
\itemize{
   \item{Ft.estimate.boot: }{a numeric array containing the estimated bootstrap variances,
   the 2.5% bootstrap quantile, and the 97.5% bootstrap quantile for all methods
   of the distribution function at all \code{p} populations. The results are shown at each \code{tval},
   \code{tval0}, \code{z.use} (if non-NULL),  \code{w.use} (if non-NULL), and for all \code{p} populations.
   The dimension of the array is \# of methods by \code{length(tval)} by \code{lenth(tval0)} by
   \code{length(z.use)} by \code{length(w.use)} by \code{p} by 3. If \code{z.use} and \code{w.use} are NULL,
   the dimension of the array is \# of methods by \code{length(tval)} by \code{lenth(tval0)}
   by \code{p} by 3.
   The last dimension
   corresponds to the bootstrap variance estimate, 2.5% bootstrap quantile,
   and the 97.5% bootstrap quantile.  The results are only valid for \eqn{t\geq t_0}, so
   \code{Ft.estimate} shows NA for any combination for which \eqn{t<t_0}.
   }

   \item {St.estimate.boot: }{a numeric array containing the estimated bootstrap variances,
   the 2.5% bootstrap quantile, and the 97.5% bootstrap quantile for all methods
   of the distribution function at all \code{m} \code{m} mixture proportion subgroups. The
   results are shown at each \code{tval}, \code{tval0}, \code{z.use} (if non-NULL), \code{w.use} (if non-NULL), and for all \code{m} mixture
   proportion subgroups.
   The dimension of the array is \# of methods by \code{length(tval)} by \code{lenth(tval0)} by
   \code{length(z.use)} by \code{length(w.use)} by \code{m} by 3.
   If \code{z.use} and \code{w.use} are NULL, then the dimension of the array is
   \# of methods by \code{length(tval)} by \code{lenth(tval0)} by \code{m} by 3
   The last dimension
   corresponds to the bootstrap variance estimate, 2.5% bootstrap quantile,
   and the 97.5% bootstrap quantile.  The results are only valid for \eqn{t\geq t_0}, so
   \code{Sout.data.out} shows NA for any combination for which \eqn{t<t_0}.
   }


   \item{Ft.AUC.BS.boot: }{a numeric array containing the estimated bootstrap variances,
   the 2.5% bootstrap quantile, and the 97.5% bootstrap quantile for all methods
   of the area under the receiver operating characteristic curve (AUC) and
   Brier Score (BS) for the \code{p} populations. The dimension of the array is \# of methods by
   \code{length(tval)} by \code{length(tval0)} by 2 by 3. The second to last stores the AUC and BS results.
  The last dimension corresponds to the bootstrap variance estimate, 2.5% bootstrap quantile,
   and the 97.5% bootstrap quantile.
   }

   \item {St.AUC.BS.boot: }{a numeric array containing the estimated bootstrap variances,
   the 2.5% bootstrap quantile, and the 97.5% bootstrap quantile for all methods
  of the area under the receiver operating characteristic curve (AUC) and
   Brier Score (BS) for the \code{m} mixture proportion subgroups. The dimension of the array is \# of methods by
   \code{length(tval)} by \code{length(tval0)} by 2 by 3. The second to last stores the AUC and BS results.
  The last dimension corresponds to the bootstrap variance estimate, 2.5% bootstrap quantile,
   and the 97.5% bootstrap quantile.
   }

}
}
\description{
Computes the bootstrap variance for all methods.
}
\examples{
# Setup parameters to generate the data
set.seed(1)
censoring.rate <- 40
p <- 2
n <- 2000
m <- 4

simu.setting <- "HD-With-Covariates"
qvs <- qvs.values(p,m)

## generate the data
data.gen <- GenerateData(n,p,m,qvs,censoring.rate,simu.setting)
x <- data.gen$x
delta <- data.gen$delta
q <- data.gen$q
ww <- data.gen$ww
zz <- data.gen$zz


## Estimation procedures to run to estimate F(t|t0,z,w)
update.qs <- FALSE
run.NPMLEs <- TRUE
run.NPNA <- TRUE
run.NPNA_avg <- FALSE
run.NPNA_wrong <- FALSE
run.OLS <- FALSE
run.WLS <- FALSE
run.EFF <- FALSE
run.EMPAVA <- FALSE


## The distribution function we are estimating is F(t|t0,z,w).
tval <- seq(0,80,by=5)  ## tval refers to "t" in F(t|t0,z,w)
tval0 <- c(0,20,30,40,50) ##tval0 refers to "t0" in F(t|t0,z,w)
z.use <- c(0,1)  ## z.use refers to "z" in  F(t|t0,z,w)
w.use <- seq(35,55,by=1)  ## w.use refers to "w" in F(t|t0,z,w)

## Setup to compute AUC/BS as in Garcia and Parast (2020). Only for simulated data.
run.prediction.accuracy <- TRUE
do_cross_validation_AUC_BS <- FALSE
know.true.groups <- TRUE
true.group.identifier <- data.gen$true.group.identifier


## Perform the estimation
estimators.out <- stride.estimator(n,m,p,qvs,q,
                                   x,delta,ww,zz,
                                   run.NPMLEs,
                                   run.NPNA,
                                   run.NPNA_avg,
                                   run.NPNA_wrong,
                                   run.OLS,
                                   run.WLS,
                                   run.EFF,
                                   run.EMPAVA,
                                   tval,tval0,
                                   z.use,w.use,
                                   update.qs,
                                   know.true.groups,
                                   true.group.identifier,
                                   run.prediction.accuracy,
                                   do_cross_validation_AUC_BS)

## Show results for the estimates
## estimators.out$Ft.estimate
## estimators.out$St.estimate

## Show results for prediction accuracy AUC and BS measures (only valid for simulated data
##  where we know the true.group.identifiers.)
## estimators.out$Ft.AUC.BS
## estimators.out$St.AUC.BS


## NOT RUN
## Do bootstrap variance
#nboot <- 100
#variance.estimation <- TRUE

#varboot <- stride.bootstrap.variance(
#						nboot,n,m,p,qvs,q,
#						x,delta,ww,zz,
#						run.NPMLEs,
#						run.NPNA,
#						run.NPNA_avg,
#						run.NPNA_wrong,
#           run.OLS,
#           run.WLS,
#           run.EFF,
#           run.EMPAVA,
#						tval,tval0,
#						z.use,w.use,
#						update.qs,
#						know.true.groups,
#						true.group.identifer,
#						estimator_Ft=estimators.out$Ft.estimate,
#						estimator_St=estimators.out$St.estimate,
#						AUC_BS_Ft=estimators.out$Ft.AUC.BS,
#						AUC_BS_St=estimators.out$St.AUC.BS,
#						run.prediction.accuracy,
#						do_cross_validation_AUC_BS=FALSE)

## Show results for the bootstrap variances of the estimates
## varboot$Ft.estimate.boot
## varboot$St.estimate.boot


## Show results for the bootstrap variances of the prediction accuracy measures, AUC and BS
## varboot$Ft.AUC.BS.boot
## varboot$St.AUC.BS.boot



}
\references{
Garcia, T.P. and Parast, L. (2020). Dynamic landmark prediction for mixture data. Biostatistics,  doi:10.1093/biostatistics/kxz052.

Garcia, T.P., Marder, K. and Wang, Y. (2017). Statistical modeling of Huntington disease onset.
In Handbook of Clinical Neurology, vol 144, 3rd Series, editors Andrew Feigin and Karen E. Anderson.

Qing, J., Garcia, T.P., Ma, Y., Tang, M.X., Marder, K., and Wang, Y. (2014).
Combining isotonic regression and EM algorithm to predict genetic risk under monotonicity constraint.
Annals of Applied Statistics, 8(2), 1182-1208.

Wang, Y., Garcia, T.P., and Ma. Y. (2012).  Nonparametric estimation for censored mixture data with
application to the Cooperative Huntington's Observational Research Trial. Journal of the American Statistical Association,
107, 1324-1338.
}
