% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stride.empava}
\alias{stride.empava}
\title{STRIDE Expectation Maximization with Pool Adjacent Violators Algorithm}
\usage{
stride.empava(n, q, x, delta, timeval, p, tol.level = 1e-05, count.max = 100)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{q}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}

\item{x}{a numeric vector of length \code{n} containing the observed event times
for each person in the sample.}

\item{delta}{a numeric vector of length \code{n} that denotes
censoring (1 denotes event is observed, 0 denotes event is censored).}

\item{timeval}{numeric value at which the distribution function is evaluated.}

\item{p}{number of populations, must be at least 2.}

\item{tol.level}{a numeric that denotes the tolerance level of the average
L1-difference of estimates of the distribution function between EM iterations. Default is 0.00001.}

\item{count.max}{a numeric that denotes the number of EM iterates to obtain. Default is 100.}
}
\value{
a list containing
\itemize{
   \item{Fest: }{numeric array of dimension \code{p} by \code{length(timeval)}
    containing EM-PAVA estimates of the distribution function evaluated at \code{timeval}.
   }
   \item{Fest.all: }{numeric array of dimension \code{p} by \code{length(timeval)+length(x)}
   containing EM-PAVA estimmates of the distribution function evaluated at all \code{timeval} and all \code{x}
   values.
   }

}
}
\description{
Wrapper function to run STRIDE estimators that computes the distribution function for the mixture data
based on an expectation-maximization (EM) algorithm that uses the pool adjacent violators algorithm (PAVA)
from isotone regression to yield a non-negative and monotone estimator. This estimator ignores covariates.
}
\references{
Garcia, T.P., Marder, K. and Wang, Y. (2017). Statistical modeling of Huntington disease onset.
In Handbook of Clinical Neurology, vol 144, 3rd Series, editors Andrew Feigin and Karen E. Anderson.

Qing, J., Garcia, T.P., Ma, Y., Tang, M.X., Marder, K., and Wang, Y. (2014).
Combining isotonic regression and EM algorithm to predict genetic risk under monotonicity constraint.
Annals of Applied Statistics, 8(2), 1182-1208.
}
