% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stride.estimator.wrapper}
\alias{stride.estimator.wrapper}
\title{Wrapper function to access STRIDE estimator}
\usage{
stride.estimator.wrapper(
  n,
  m,
  p,
  qvs,
  q,
  x,
  delta,
  ww,
  zz,
  run.NPMLEs,
  run.NPNA,
  run.NPNA_avg,
  run.NPNA_wrong,
  run.OLS,
  run.WLS,
  run.EFF,
  run.EMPAVA,
  tval,
  tval0,
  z.use,
  w.use,
  update.qs = FALSE,
  know.true.groups = FALSE,
  true.group.identifier = NULL,
  run.prediction.accuracy = FALSE,
  do_cross_validation_AUC_BS = FALSE
)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{p}{number of populations, must be at least 2.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{q}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}

\item{x}{a numeric vector of length \code{n} containing the observed event times
for each person in the sample.}

\item{delta}{a numeric vector of length \code{n} that denotes
censoring (1 denotes event is observed, 0 denotes event is censored).}

\item{ww}{a numeric vector of length \code{n} containing the values of the continuous
covariate for each person in the sample. Can be NULL.}

\item{zz}{a numeric vector of length \code{n} containing the values of the discrete
covariate for each person in the sample. Can be NULL.}

\item{run.NPMLEs}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data based on the type-I and type II nonparametric maximum likelihood
estimators. The type I nonparametric maximum likelihood estimator is referred
to as the "Kaplan-Meier" estimator in Garcia and Parast (2020). Neither the type I nor type II
adjust for covariates.}

\item{run.NPNA}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that accounts for covariates and dynamic
landmarking. This estimator is called "NPNA" in Garcia and Parast (2020).}

\item{run.NPNA_avg}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that averages out over the observed covariates.
This is referred to as NPNA_marg in Garcia and Parast (2020).}

\item{run.NPNA_wrong}{a logical indicator. If TRUE, then the output includes the
estimated distribution function for mixture data that adjusts for covariates, but
ignores landmarking. This is referred to as NPNA_{t_0=0} in Garcia and Parast (2020).}

\item{run.OLS}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using an ordinary least squares influence function. The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.WLS}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using a weighted least squares influence function. The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.EFF}{a logical indicator. If TRUE, then the output includes the estimated distribution function
computed using the efficient influence function based on Hilbert space projection theory results.
The estimator adjusts for censoring using
inverse probability weighting (IPW), augmented inverse probability weighting (AIPW),
and imputation (IMP). See details in Wang et al (2012). These estimators do not
adjust for covariates.}

\item{run.EMPAVA}{logical indicator. If TRUE, we compute the distribution function for the mixture data
based on an expectation-maximization (EM) algorithm that uses the pool adjacent violators algorithm (PAVA)
from isotone regression to yield a non-negative and monotone estimator. This estimator does not adjust
for covariates.  See details in Qing et al (2014).}

\item{tval}{numeric vector of time points at which the distribution function is evaluated, all values must
be non-negative.}

\item{tval0}{numeric vector of time points representing the landmark times. All values must be non-negative
and smaller than the maximum of \code{tval}.}

\item{z.use}{numeric vector at which to evaluate the discrete covariate \eqn{Z} at in the estimated distribution function.
The values of \code{z.use} must be in the range of the observed \code{zz}.  Can be NULL.}

\item{w.use}{numeric vector at which to evaluate the continuous covariate \eqn{W} at in the estimated distribution function.
The values of \code{w.use} must be in the range of the observed \code{ww}.  Can be NULL.}

\item{update.qs}{logical indicator. If TRUE, the mixture proportions \code{q} will be updated.
This is currently not implemented.}

\item{know.true.groups}{logical indicator. If TRUE, then we know the population identifier for each person in the sample.
This option is only used for simulation studies to check prediction accuracy. Default is FALSE.}

\item{true.group.identifier}{numeric vector of length \code{n} denoting the
population identifier for each person in the sample.
Default is NULL.}

\item{run.prediction.accuracy}{logical indicator. If TRUE, then we compute the prediction accuracy measures, including the
area under the receiver operating characteristic curve (AUC) and the Brier Score (BS). Prediction accuracy is only valid
in simulation studies where \code{know.true.groups}=TRUE and \code{true.group.identifier} is available.}

\item{do_cross_validation_AUC_BS}{logical indicator. If TRUE, then we compute the prediction accuracy measures, including the
area under the receiver operating characteristic curve (AUC) and the Brier Score (BS) using cross-validation. Prediction accuracy is only valid
in simulation studies where \code{know.true.groups}=TRUE and \code{true.group.identifier} is available.}
}
\value{
\code{stride.wrapper.estimator} returns a list containing
\itemize{
   \item{problem: }{a numeric indicator of errors in the NPNA, NPNA_avg, NPNA_wrong estimator. If NULL, no error is reported.
   Otherwise, there is an error in the computation of the NPNA, NPNA_avg, or NPNA_wrong estimator.}

   \item{Ft.estimate: }{a numeric array containing the estimated distribution functions for all methods for all
   \code{p} populations. The distribution function is evaluated at each \code{tval},
   \code{tval0}, \code{z.use} (if non-NULL), \code{w.use} (if non-NULL), and for all \code{p} populations.
   The dimension of the array is \# of methods by \code{length(tval)} by \code{lenth(tval0)} by
   \code{length(z.use)} by \code{length(w.use)} by \code{p}. If \code{z.use} and \code{w.use} are NULL,
   the dimension of the array is  \# of methods by \code{length(tval)} by \code{lenth(tval0)} by \code{p}. The distribution function is only valid for \eqn{t\geq t_0}, so
   \code{Ft.estimate} shows NA for any combination for which \eqn{t<t_0}.
   }

   \item {St.estimate: }{a numeric array containing the estimated distribution functions for all methods
   for all \code{m} mixture proportion subgroups. The distribution function is evaluated
   at each \code{tval}, \code{tval0}, \code{z.use} (if non-NULL), \code{w.use} (if non-NULL), and for all \code{m} mixture
   proportion subgroups.
   The dimension of the array is \# of methods by \code{length(tval)} by \code{lenth(tval0)} by
   \code{length(z.use)} by \code{length(w.use)} by \code{m}. If \code{z.use} and \code{w.use} are NULL,
   the dimension of the array   is \# of methods by \code{length(tval)} by \code{lenth(tval0)} by \code{m}.  The distribution function is only valid for \eqn{t\geq t_0}, so
   \code{St.estimate} shows NA for any combination for which \eqn{t<t_0}.
   }

   \item{Ft.AUC.BS: }{a numeric array containing the
   area under the receiver operating characteristic curve (AUC) and
   Brier Score (BS) for the \code{p} populations. The dimension of the array is \# of methods by
   \code{length(tval)} by \code{length(tval0)} by 2,
   where the last dimension stores the AUC and BS results.

   Results for both the estimated distributon functions and prediction
   accuracy measures (AUC, BS) are only valid when \eqn{t\geq t_0}, so
   arrays show NA for any combination for which \eqn{t<t_0}.

   }

   \item {St.AUC.BS: }{a numeric array containing the results are the
   area under the receiver operating characteristic curve (AUC) and
   Brier Score (BS) for the \code{m} mixture proportion groups.
   The dimension of the array is \# of methods by
   \code{length(tval)} by \code{length(tval0)} by 2,
   where the last dimension stores the AUC and BS results.

   Results for both the estimated distributon functions and prediction
   accuracy measures (AUC, BS) are only valid when \eqn{t\geq t_0}, so
   arrays show NA for any combination for which \eqn{t<t_0}.

   }

}
}
\description{
Checks for common errors in user input, assembles data user provides and applies STRIDE estimators.
}
\section{Details}{

We estimate nonparametric distribution functions for mixture data  where
the population identifiers are unknown, and the probability of belonging
to a population is known (typically estimated with external data).
The distribution functions are evaluated at
time points \code{tval}. All estimators adjust for dynamic landmark prediction.
Dynamic landmark prediction means that the distribution function is computed knowing
that the survival time, \eqn{T}, satisfies \eqn{T >t_0}
where \eqn{t_0} are the time points in \code{tval0}. The NPNA, NPNA_avg,
and NPNA_wrog adjust for one discrete covariate (\code{zz}) and one continuous covariate (\code{ww}).
}

\references{
Garcia, T.P. and Parast, L. (2020). Dynamic landmark prediction for mixture data. Biostatistics,  doi:10.1093/biostatistics/kxz052.

Garcia, T.P., Marder, K. and Wang, Y. (2017). Statistical modeling of Huntington disease onset.
In Handbook of Clinical Neurology, vol 144, 3rd Series, editors Andrew Feigin and Karen E. Anderson.

Qing, J., Garcia, T.P., Ma, Y., Tang, M.X., Marder, K., and Wang, Y. (2014).
Combining isotonic regression and EM algorithm to predict genetic risk under monotonicity constraint.
Annals of Applied Statistics, 8(2), 1182-1208.

Wang, Y., Garcia, T.P., and Ma. Y. (2012).  Nonparametric estimation for censored mixture data with
application to the Cooperative Huntington's Observational Research Trial. Journal of the American Statistical Association,
107, 1324-1338.
}
