% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{stride.nocovariates}
\alias{stride.nocovariates}
\title{STRIDE estimators without covariates}
\usage{
stride.nocovariates(
  n,
  q,
  x,
  delta,
  timeval,
  qvs,
  p,
  m,
  r,
  boot,
  bootvar,
  useOLS,
  useWLS,
  useEFF,
  useNPMLEs,
  useEMPAVA
)
}
\arguments{
\item{n}{sample size, must be at least 1.}

\item{q}{a numeric matrix of size \code{p} by \code{n} containing the
mixture proportions for each person in the sample.}

\item{x}{a numeric vector of length \code{n} containing the observed event times
for each person in the sample.}

\item{delta}{a numeric vector of length \code{n} that denotes
censoring (1 denotes event is observed, 0 denotes event is censored).}

\item{timeval}{numeric value at which the distribution function is evaluated.}

\item{qvs}{a numeric matrix of size \code{p} by \code{m} containing all possible
mixture proportions (i.e., the probability of belonging to each population k, k=1,...,p.).}

\item{p}{number of populations, must be at least 2.}

\item{m}{number of different mixture proportions, must be at least 2.}

\item{r}{numeric vector including the number of individuals in each mixture proportion
group.}

\item{boot}{number of bootstrap replicates.}

\item{bootvar}{logical indicator. If TRUE, we compute the bootstrap variance estimates of
the estimators.}

\item{useOLS}{logical indicator. If TRUE, we compute the distribution function
for the mixture data where the influence function is the ordinary least squares estimator.}

\item{useWLS}{logical indicator. If TRUE, we compute the distribution function
for the mixture data where the influence function is the weighted least squares estimnator.}

\item{useEFF}{logical indicator. If TRUE, we compute the distribution function
for the mixture data where the influence function is the efficient estimator.}

\item{useNPMLEs}{logical indicator. If TRUE, we compute the distribution function
for the mixture data based on the type I and type II
nonparametric maximum likelihood esimators (NPMLEs).}

\item{useEMPAVA}{logical indicator. If TRUE, we compute the distribution function for the mixture data
based on an expectation-maximization (EM) algorithm that uses the pool adjacent violators algorithm (PAVA)
from isotone regression to yield a non-negative and monotone estimator.}
}
\value{
a list containing
\itemize{
   \item{hts0: }{numeric vector of length \code{m} containing the
   Kaplan-Meier estimates of the survival curves for the \code{m} mixture
   proportion groups.}
   \item{Fest: }{numeric array containing the estimated distribution functions for
   all influence functions of interest.}
   \item{var_est: }{numeric array containing the estimated variances for
   all influence functions of interest.}
   \item{eflag: }{numeric value indicating if errors occurred in computing  the estimated distribution functions.
   \code{eflag}<0 indicates an error.}

}
}
\description{
Wrapper function to run STRIDE estimators that
ignore covariates.
}
\references{
Garcia, T.P., Marder, K. and Wang, Y. (2017). Statistical modeling of Huntington disease onset.
In Handbook of Clinical Neurology, vol 144, 3rd Series, editors Andrew Feigin and Karen E. Anderson.

Qing, J., Garcia, T.P., Ma, Y., Tang, M.X., Marder, K., and Wang, Y. (2014).
Combining isotonic regression and EM algorithm to predict genetic risk under monotonicity constraint.
Annals of Applied Statistics, 8(2), 1182-1208.

Wang, Y., Garcia, T.P., and Ma. Y. (2012).  Nonparametric estimation for censored mixture data with
application to the Cooperative Huntington's Observational Research Trial. Journal of the American Statistical Association,
107, 1324-1338.
}
