% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stride.R
\docType{package}
\name{stride}
\alias{stride}
\title{stride package}
\description{
This package provides functions to nonparametrically estimate survival models
from mixture data. Mixture data is when data are collected from multiple populations
of interest, but no information is provided as to which population each observation belongs.
The functions are for nonparametric prediction models that incorporate covariates and landmarking.
}
\references{
Garcia, T.P. and Parast, L. (2020). Dynamic landmark prediction for mixture data. Biostatistics,  doi:10.1093/biostatistics/kxz052.

Garcia, T.P., Marder, K. and Wang, Y. (2017). Statistical modeling of Huntington disease onset.
In Handbook of Clinical Neurology, vol 144, 3rd Series, editors Andrew Feigin and Karen E. Anderson.

Qing, J., Garcia, T.P., Ma, Y., Tang, M.X., Marder, K., and Wang, Y. (2014).
Combining isotonic regression and EM algorithm to predict genetic risk under monotonicity constraint.
Annals of Applied Statistics, 8(2), 1182-1208.

Wang, Y., Garcia, T.P., and Ma. Y. (2012).  Nonparametric estimation for censored mixture data with
application to the Cooperative Huntington's Observational Research Trial. Journal of the American Statistical Association,
107, 1324-1338.
}
